Part-2/backend/requirements.txt
fastapi
pydantic
langchain_pinecone
langchain_openai
langchain-community
python-dotenv
uvicorn